probando render

+++ src/main/java/com/github/jcapitanmoreno/beenice_api/repositories/GastoRespository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GastoRespository extends JpaRepository<Gasto, Integer> {
+public interface GastoRespository extends JpaRepository<Gasto, Long> {
+
+
 
 
 }
